requirements.txt
langchain-openai
langchain
geocoder
langgraph
keras
tensorflow
langchain_chroma
langchain_community
grandalf
langchain_upstage
certifi
trulens
trulens
trulens-eval
trulens-core
trulens-apps-langchain
trulens-providers-langchain
trulens-providers-openai
azure-search-documents==11.4.0
azure-identity
python-dotenv
pytest
folium
requests
flasgger
Pillow
Flask==2.3.2
gunicorn==20.1.0
